pyproject.toml
[project]
name = "cacten"
version = "0.1.0"
description = "Personalized RAG middleware layer that wraps Claude Code via MCP"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "langchain-text-splitters>=1.1.1",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "pypdf>=6.9.1",
    "qdrant-client>=1.17.1",
    "rank-bm25>=0.2.2",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[project.scripts]
cacten = "cacten.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cacten"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.12"
mypy_path = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
